feat(page.tsx): redesign SchoolsPage layout to separate assigned school from others fix(page.tsx): correct typo in invisible text from 'Shool Detials' to 'School Details' feat(header.tsx): add border color to header feat(icons.tsx): add new edit icon feat(navigation.tsx): rearrange navigation items and adjust icon sizes feat(tailwind.config.ts): add new color 'shamiri-light-blue' to tailwind config
The changes were made to improve the user interface and user experience by making the layout more intuitive and visually appealing. The navigation items were rearranged for better flow, and a new color was added to the tailwind config for more design options. An edit icon was also added for future use.